+++ RansacBot.Net5.0/HystoryTest/Form2(old).cs
@@ -46,7 +46,7 @@ namespace RansacBot.HystoryTest
 			feeder.Subscribe(new("", ""), 
 				(tick) =>
 					{
-						tradesHystory.CheckForStops((decimal)tick.PRICE);
+						tradesHystory.CheckForStops(tick.PRICE);
 					});
 
 			monkeyNFinder.NewExtremum += higherLowerFilter.OnNewExtremum;
@@ -66,7 +66,7 @@ namespace RansacBot.HystoryTest
 
 			tradesHystory.ExecutedLongStop += (price, exPrice) =>
 			{
-				int tradeIndex = longs.FindIndex((trade) => (decimal)trade.trade.stop.price == price);
+				int tradeIndex = longs.FindIndex((trade) => trade.trade.stop.price == price);
 				TradeWithStopWithTick trade = longs[tradeIndex];
 				longs.RemoveAt(tradeIndex);
 				dealsWriter.WriteLine(
@@ -81,7 +81,7 @@ namespace RansacBot.HystoryTest
 			};
 			tradesHystory.ExecutedShortStop += (price, exPrice) =>
 			{
-				int tradeIndex = shorts.FindIndex((trade) => (decimal)trade.trade.stop.price == price);
+				int tradeIndex = shorts.FindIndex((trade) => trade.trade.stop.price == price);
 				TradeWithStopWithTick trade = shorts[tradeIndex];
 				shorts.RemoveAt(tradeIndex);
 				dealsWriter.WriteLine(
@@ -96,7 +96,7 @@ namespace RansacBot.HystoryTest
 			};
 			tradesHystory.KilledLongStop += (price, exPrice) =>
 			{
-				int tradeIndex = longs.FindIndex((trade) => (decimal)trade.trade.stop.price == price);
+				int tradeIndex = longs.FindIndex((trade) => trade.trade.stop.price == price);
 				TradeWithStopWithTick trade = longs[tradeIndex];
 				longs.RemoveAt(tradeIndex);
 				dealsWriter.WriteLine(
@@ -111,7 +111,7 @@ namespace RansacBot.HystoryTest
 			};
 			tradesHystory.KilledShortStop += (price, exPrice) =>
 			{
-				int tradeIndex = shorts.FindIndex((trade) => (decimal)trade.trade.stop.price == price);
+				int tradeIndex = shorts.FindIndex((trade) => trade.trade.stop.price == price);
 				TradeWithStopWithTick trade = shorts[tradeIndex];
 				shorts.RemoveAt(tradeIndex);
 				dealsWriter.WriteLine(
